Move precharge fns to precharge.rs
